Added validation for column names. Minor refactoring.
