+++ src/main/java/edu/pitt/dbmi/data/validation/MessageType.java
@@ -30,7 +30,7 @@ public enum MessageType {
     FILE_MISSING_VALUE,
     FILE_INVALID_NUMBER,
     FILE_EXCESS_DATA,
-    FILE_INSUFFICIENT_DAT,
+    FILE_INSUFFICIENT_DATA,
     FILE_SUMMARY
 
 }
