+++ src/test/resources/data/tabular/continuous/validation/sim_test_data.csv
@@ -1,4 +1,4 @@
-X1,X2,X3,X4,X5,X6,X7,X8,X9,X10
+X1,X2,X3,X4,X5,X6,,X8,"X9,X_9",X10
 0.5148,2.5901,-0.0721,1.4016,0.0888,-7.9602,1.1406,4.4783,1.8136,-1.9697
 0.0612,-3.68,4.0955,1.2874,-0.801,4.1415,-0.7477,-0.5047,0.6948,0.5274
 -1.4067,0.1086,3.1234,-0.6666,-0.5165,3.4851,1.2906,1.1263,0.3555,-0.8132
